refactor: organization of login-related actions
